Send multiple terraform resources to templates
In the past terrible would only send the first terraform resource to
jinja templates. This fix sends all of the resources to jinja templets.
This is a breaking change that will require all templates written
against previous versions of terrible to accept a list of resources.
This can be accomplished simply by adding the following to a template
{% for resource in resources -%}
… normal template logic
{%- endfor %}

Also linted some extra whitespace.